Porting more EM tests to behaviors
